requirements.txt
discord.py>=2.3.0
gspread>=5.12.0
google-auth>=2.25.0
python-dotenv>=1.0.0
apscheduler>=3.10.4
pytz>=2024.1
